# Optimize retrieving indexed field values