--- a/fiftyone/core/collections.py
+++ b/fiftyone/core/collections.py
@@ -9180,10 +9180,14 @@ class SampleCollection(object):
                         field,
                         values_only=True,
                     )
-                except Exception as e:
+                except ValueError as e:
                     # Fallback to values, but log the error to surface the
                     # recommendation to create an index
-                    logger.warning(e)
+                    logger.debug(e)
+                except Exception as e:
+                    # Unexpected error, so we should surface it in the logs
+                    # but still fallback to values
+                    logger.error(e)
 
         make = lambda field_or_expr: foa.Values(
             field_or_expr,