## What changes are proposed in this pull request?

- force mongo to use a covered index query when possible to improve speed, particularly on large datasets

- context/background: aggregation projections can never be a fully covered query and will always involve a collection scan. By using find + hint, we force the lookup on the index only, vastly improving performance

- We can build upon this iteratively and expand the supported fields, but this will reduce the bottleneck in a lot of existing workflows. 


## How is this patch tested? If it is not, please explain why.

- added unit tests
- tested by running locally:

### values() optimization
```
In [16]: %%time
    ...: values=ds.values("_id")
    ...: 
    ...: 
CPU times: user 17.5 s, sys: 4.88 s, total: 22.4 s
Wall time: 7min 3s

In [17]: %%time
    ...: values=ds.values("_id", _enforce_natural_order=False)
    ...: 
    ...: 
CPU times: user 8.07 s, sys: 1.68 s, total: 9.75 s
Wall time: 19.7 s

```

```
In [19]: ds=fo.load_dataset('big-image-ds')

In [20]: %%time
    ...: vals=ds.values("filepath")
    ...: 
    ...: 
CPU times: user 3min 24s, sys: 1min 20s, total: 4min 45s
Wall time: 2h 9min 14s

In [21]: len(vals)
Out[21]: 32877680


In [25]: %%time
    ...: vals=ds.values("filepath", _enforce_natural_order=False)
    ...: 
    ...: 
CPU times: user 1min 24s, sys: 10.3 s, total: 1min 34s
Wall time: 4min 32s

In [26]: len(vals)
Out[26]: 32877680

```

### Optimize getting values of a compound index
```
In [14]: %%time
    ...: mapping=get_indexed_values(ds, "_id_1_filepath_1")
    ...: 
    ...: 
CPU times: user 9.57 s, sys: 1.86 s, total: 11.4 s
Wall time: 45.4 s

In [15]: %%time
    ...: values=ds.values(["_id","filepath"])
    ...: 
    ...: 
CPU times: user 18.6 s, sys: 3.36 s, total: 22 s
Wall time: 10min 38s
```

## Release Notes

### Is this a user-facing change that should be mentioned in the release notes?

<!--
Please fill in relevant options below with an "x", or by clicking the checkboxes
after submitting this pull request. Example:
-   [x] Selected option
-->

-   [ ] No. You can skip the rest of this section.
-   [x] Yes. Give a description of this change to be included in the release
        notes for FiftyOne users.

Add a new function to retrieve values from indexed fields fast. 
Enable using this optimization on full datasets for values() calls by passing `_enforce_natural_order=False`

### What areas of FiftyOne does this PR affect?

-   [ ] App: FiftyOne application changes
-   [ ] Build: Build and test infrastructure changes
-   [x] Core: Core `fiftyone` Python library changes
-   [ ] Documentation: FiftyOne documentation changes
-   [ ] Other


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **Performance Improvements**
  - Optimized retrieval of indexed field values for full datasets, speeding up queries on fields like `id`, `_id`, and `filepath`.
- **New Features**
  - Introduced efficient methods to fetch values from indexed fields, improving query performance.
  - Exposed new indexed value retrieval functions in the public API.
- **Bug Fixes**
  - Enhanced error handling with clear errors for queries on non-indexed or nonexistent fields.
- **Tests**
  - Added tests validating indexed value retrieval behavior and error handling.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->