@kaixi-wang 2025-04-18
> Wow I'm impressed! Mongo DB can be so silly sometimes. How could it not decide that that find query should use _id index ... Note that you can also force it by putting something dumb in the query slot.
> 
> Unfortunately we cannot merge as is. This will totally break the contract (even if unofficial) we have maintained for `values()`. They need to be in the same order. For example in `set_values()`, using this form of `values('id')` will give us a potentially different ordering, causing values to be assigned in the wrong order.
> 
> This would have to be part of a different path: _id sorted samples which we have discussed before. It would have to be opt-in and only useable in certain situations. What could use this initially though, map_samples()? Faster paging than $skip/$limit?

yeah valid points. They key thing here is that for this level of speedup, we just can't use aggregate() because it will always end in a doc lookup rather than relying entirely on the index