@brimoor 2025-04-18
> What could use this initially though, map_samples()?

+100 to this suggestion!