@benjaminpkane 2025-04-18
With a compound index, i.e. `(_id, filepath)`, order should be maintained. See https://github.com/voxel51/fiftyone/commit/61e4ed06cc2861f54a7b8d57c723e6e6f3f8d0b5