@swheaton 2025-04-18
> With a compound index, i.e. `(_id, filepath)`, order should be maintained. See [61e4ed0](https://github.com/voxel51/fiftyone/commit/61e4ed06cc2861f54a7b8d57c723e6e6f3f8d0b5)

How's that possible? Regular values() without a sort returns in insertion order. If we force it to use the index, it will return it in sorted order.