@swheaton 2025-04-18
In that case of the server you're attaching ID to the result so it's fine