@swheaton 2025-04-18
I think you are right that values('filepath') should be in _id order if it's using that compound index.
But _id order is not necessarily default/insertion order