@benjaminpkane 2025-04-18
Ok, I don't know the use case details. Yes, the compound index ensures the value pairs are a correct. And the results will follow the `_id` order due to the index