@swheaton 2025-04-18
> Ok, I don't know the use case details. Yes, the compound index ensures the value pairs are a correct. And the results will follow the `_id` order due to the index

Thanks though. We can probably use it in the future if we are able to relax some constraints or create new use case pathways